+++ crates/cli/src/commands/apply_pattern.rs
@@ -29,6 +29,7 @@ use std::sync::atomic::Ordering;
 use std::collections::BTreeMap;
 use tokio::fs;
 
+use crate::diff::extract_target_ranges;
 use crate::{
     analyze::par_apply_pattern,
     community::parse_eslint_output,
@@ -48,6 +49,7 @@ use marzano_messenger::{
 use crate::resolver::{get_grit_files_from_cwd, GritModuleResolver};
 use crate::utils::has_uncommitted_changes;
 
+use super::apply::SharedApplyArgs;
 use super::init::init_config_from_cwd;
 
 #[derive(Deserialize)]
@@ -103,13 +105,6 @@ pub struct ApplyPatternArgs {
         )]
     pub visibility: VisibilityLevels,
     #[clap(
-        long = "only-in-diff",
-        help = "Only rewrite ranges that are inside the unified diff if a path to the diff is provided, or the results of git diff HEAD if no path is provided.",
-        hide = true,
-        conflicts_with = "only_in_json"
-    )]
-    only_in_diff: Option<Option<PathBuf>>,
-    #[clap(
         long = "only-in-json",
         help = "Only rewrite ranges that are inside the provided eslint-style JSON file",
         hide = true,
@@ -146,7 +141,6 @@ impl Default for ApplyPatternArgs {
             format: Default::default(),
             interactive: Default::default(),
             visibility: VisibilityLevels::Hidden,
-            only_in_diff: Default::default(),
             only_in_json: Default::default(),
             output_file: Default::default(),
             cache: Default::default(),
@@ -173,6 +167,7 @@ macro_rules! flushable_unwrap {
 #[allow(clippy::too_many_arguments, unused_mut)]
 pub(crate) async fn run_apply_pattern(
     mut pattern: String,
+    shared: SharedApplyArgs,
     paths: Vec<PathBuf>,
     arg: ApplyPatternArgs,
     multi: MultiProgress,
@@ -227,15 +222,8 @@ pub(crate) async fn run_apply_pattern(
     let filter_range = if let Some(json_path) = arg.only_in_json.clone() {
         let json_ranges = flushable_unwrap!(emitter, parse_eslint_output(json_path));
         Some(json_ranges)
-    } else if let Some(Some(diff_path)) = &arg.only_in_diff {
-        let diff_ranges = flushable_unwrap!(emitter, extract_modified_ranges(diff_path));
-        Some(diff_ranges)
-    } else if let Some(None) = &arg.only_in_diff {
-        let diff = git_diff(&std::env::current_dir()?)?;
-        let diff_ranges = flushable_unwrap!(emitter, parse_modified_ranges(&diff));
-        Some(diff_ranges)
     } else {
-        None
+        flushable_unwrap!(emitter, extract_target_ranges(&shared))
     };
 
     let (my_input, lang) = if let Some(pattern_libs) = pattern_libs {
